Refactor code to prevent issue with GCC 7.5
Still not sure if this is an issue with GCC or with GeCo3/GeDe3.
We can only trigger this in GCC 7.5 with -Ofast -march=native.
Clang 6 and 10 and GCC 10 work fine.
Valgrind and sanitizers don't show any errors.
The issue manifests as a floating point difference in the neural network calculations.
The calculations done during compression and decompression begin to diverge by less than 0.000001.
Because the results of the network are used as inputs to the network the errors start to accumulate.